+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@egjs/flicking-plugins": "^4.7.1",
     "@egjs/react-flicking": "^4.11.0",
+    "@hookform/resolvers": "^3.3.1",
     "@supabase/supabase-js": "^2.32.0",
     "@tanstack/react-query": "^4.32.6",
     "@testing-library/jest-dom": "^5.14.1",
@@ -34,6 +35,7 @@
     "tinycolor2": "^1.6.0",
     "typescript": "*",
     "web-vitals": "^2.1.0",
+    "yup": "^1.2.0",
     "zustand": "^4.4.1"
   },
   "scripts": {
