+++ src/pages/UpdatePassword.tsx
@@ -2,44 +2,31 @@ import { type SubmitHandler, useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
 import toast from "react-simple-toasts";
 
+import { yupResolver } from "@hookform/resolvers/yup";
 import { changePassword } from "api/supabase/auth";
 import { InvalidText } from "components";
+import { updateUserPasswordSchema } from "schema/formSchema";
 
 interface UpdatePasswordInput {
-  newPassword: string;
-  newPasswordConfirm: string;
+  password: string;
+  passwordConfirm: string;
 }
 
 const UpdatePassword = () => {
   const navigate = useNavigate();
-  const {
-    register,
-    handleSubmit,
-    getValues,
-    formState: { errors },
-  } = useForm<UpdatePasswordInput>();
+
+  const resolver = yupResolver(updateUserPasswordSchema);
+  const { register, handleSubmit, formState } = useForm<UpdatePasswordInput>({ resolver });
+  const { errors } = formState;
 
   const onSubmit: SubmitHandler<UpdatePasswordInput> = async (data) => {
-    const { newPassword } = data;
+    const { password } = data;
 
     try {
-      await changePassword(newPassword);
-      toast("비밀번호가 정상적으로 변경되었습니다.", { theme: "plain", zIndex: 9999 });
+      await changePassword(password);
+      toast("비밀번호가 정상적으로 변경되었습니다.", { theme: "warning", zIndex: 9999 });
       navigate("/");
-    } catch (error) {
-      switch (error) {
-        case "New password should be different from the old password.":
-          toast("이전 비밀번호와 동일합니다.", { theme: "failure", zIndex: 9999 });
-          break;
-        case "Auth session missing!":
-          toast("이메일 유효시간이 만료되었습니다.", { theme: "failure", zIndex: 9999 });
-          break;
-        default:
-          toast("비밀번호 변경에 실패하였습니다.", { theme: "failure", zIndex: 9999 });
-          console.error("newError : ", error);
-          break;
-      }
-    }
+    } catch (error) {}
   };
 
   return (
@@ -53,48 +40,33 @@ const UpdatePassword = () => {
           <p>새로운 비밀번호를 등록 후 사용해주세요.</p>
         </div>
         <form onSubmit={handleSubmit(onSubmit)} action="#" className="w-full mt-10 flex-column">
-          <label htmlFor="newPassword" className="self-start body-4 my-[8px]">
+          <label htmlFor="password" className="self-start body-4 my-[8px]">
             새 비밀번호
           </label>
           <div className="flex w-full">
             <input
-              id="newPassword"
+              id="password"
               type="password"
               placeholder="새 비밀번호"
               className="auth-input body-3"
-              {...register("newPassword", {
-                required: "비밀번호를 입력해주세요.",
-                pattern: {
-                  value: /^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&])[A-Za-z\d@$!%*#?&]{8,}$/,
-                  message: "영문 대문자, 영문 소문자, 숫자, 특수문자가 하나 이상 포함되어야 합니다.",
-                },
-                minLength: { value: 6, message: "비밀번호가 너무 짧습니다." },
-                validate: {
-                  matchesPreviousPassword: (value) => {
-                    const prevPassword = getValues("newPasswordConfirm");
-                    return prevPassword === value || "비밀번호가 일치하지 않습니다.";
-                  },
-                },
-              })}
+              {...register("password")}
             />
           </div>
-          <InvalidText errorsMessage={errors.newPassword?.message} />
+          <InvalidText errorsMessage={errors.password?.message} />
 
-          <label htmlFor="newPasswordConfirm" className="self-start my-2 body-4">
+          <label htmlFor="passwordConfirm" className="self-start my-2 body-4">
             새 비밀번호 확인
           </label>
           <div className="flex w-full">
             <input
-              id="newPasswordConfirm"
+              id="passwordConfirm"
               type="password"
               placeholder="새 비밀번호 확인"
               className="auth-input body-3"
-              {...register("newPasswordConfirm", {
-                required: "비밀번호를 입력해주세요.",
-              })}
+              {...register("passwordConfirm")}
             />
           </div>
-          <InvalidText errorsMessage={errors.newPasswordConfirm?.message} />
+          <InvalidText errorsMessage={errors.passwordConfirm?.message} />
           <button className="auth-button point-button">변경하기</button>
         </form>
       </div>
