--- src/pages/mypage/UpdateUser.tsx
@@ -1,373 +1,0 @@
-import { useState } from "react";
-import { type SubmitHandler, useForm } from "react-hook-form";
-import { useNavigate } from "react-router-dom";
-import toast from "react-simple-toasts";
-import uuid from "react-uuid";
-
-import {
-  changeMetaAvatar,
-  changeMetaName,
-  changePassword,
-  deleteImage,
-  deleteUser,
-  fetchUserCheckData,
-  logout,
-  uploadImage,
-} from "api/supabase/auth";
-import { STORAGE_URL } from "api/supabase/supabaseClient";
-import defaultImg from "assets/defaultImg.jpg";
-import defaultImgWebp from "assets/defaultImgWebp.webp";
-import photoCamera from "assets/svgs/photoCamera.svg";
-import xmark from "assets/svgs/xmark.svg";
-import {
-  PasswordVisibleButton,
-  InvalidText,
-  Title,
-  passwordValid,
-  nameValid,
-  useDialog,
-  DeleteConfirmText,
-} from "components";
-import { useAuthQuery } from "hooks/useAuthQuery";
-import { useImageResize } from "hooks/useImageResize";
-import Error from "pages/Error";
-import { useAuthStore } from "store";
-
-import type { PasswordVisible } from "components";
-
-interface UpdateInput {
-  name: string;
-  password: string;
-  passwordConfirm: string;
-}
-
-const LABEL_STYLE = "self-center w-[136px] px-[24px] text-[14px] font-normal leading-[130%]";
-
-const UpdateUser = () => {
-  const navigate = useNavigate();
-  const { Confirm } = useDialog();
-
-  const [showPassword, setShowPassword] = useState<PasswordVisible>({ password: false, passwordConfirm: false });
-  const [checkedDuplicate, setCheckedDuplicate] = useState(false);
-  const [isOpenToggle, setIsOpenToggle] = useState({ name: false, password: false });
-  const { imageFile, setImageFile, resizePixelHandler, imageSizeSaveHandler, resizeFile } = useImageResize();
-
-  const { currentUserResponse, patchUserMutation } = useAuthQuery();
-  const { data: currentUser } = currentUserResponse;
-  const { currentUserId } = useAuthStore();
-
-  if (currentUser === undefined || currentUserId === undefined) {
-    navigate("/");
-    return <Error />;
-  }
-
-  const { name: currentName, avatar_url: currentProfileImg } = currentUser;
-  const userId = currentUserId;
-
-  const prevProfileImageId =
-    currentProfileImg === "" ? "" : currentProfileImg.replace(`${STORAGE_URL}/profileImg/`, "");
-
-  const {
-    register,
-    handleSubmit,
-    getValues,
-    setError,
-    resetField,
-    formState: { errors },
-  } = useForm<UpdateInput>();
-
-  const submitProfileImgHandler: SubmitHandler<UpdateInput> = async () => {
-    const uid = uuid();
-    if (imageFile === null) return;
-
-    const resizePixel = await resizePixelHandler(160);
-    const resizeImageFile = await resizeFile(imageFile, resizePixel);
-    if (resizeImageFile === undefined) return;
-
-    const profileImg = `${STORAGE_URL}/profileImg/${uid}`;
-    await deleteImage(prevProfileImageId);
-    patchUserMutation.mutate({ inputValue: { avatar_url: profileImg }, userId });
-    await changeMetaAvatar(profileImg);
-    void uploadImage({ file: resizeImageFile, userId: uid });
-    setImageFile(null);
-  };
-
-  const resetImgFile = async () => {
-    if (imageFile !== null) {
-      setImageFile(null);
-    }
-    if (currentProfileImg !== "") {
-      await deleteImage(prevProfileImageId);
-      await changeMetaAvatar("");
-      patchUserMutation.mutate({ inputValue: { name: currentName, avatar_url: "" }, userId });
-    }
-  };
-
-  const duplicateCheck = async () => {
-    const getUserData = await fetchUserCheckData();
-
-    const matchUser = getUserData.filter((user) => user.name === getValues("name"));
-
-    if (matchUser === null || matchUser.length === 0) setCheckedDuplicate(true);
-    else setError("name", { message: "이미 존재하는 닉네임입니다." });
-  };
-
-  const toggleChangeHandler = (target: "name" | "password") => {
-    if (target === "name") {
-      setIsOpenToggle({ password: false, name: !isOpenToggle.name });
-      resetField("name");
-    } else {
-      setIsOpenToggle({ name: false, password: !isOpenToggle.password });
-      resetField("password");
-      resetField("passwordConfirm");
-    }
-  };
-
-  const changePasswordHandler: SubmitHandler<UpdateInput> = async (data) => {
-    try {
-      await changePassword(data.password);
-      toast("비밀번호가 정상적으로 변경되었습니다.", { theme: "warning", zIndex: 9999 });
-    } catch (error) {}
-
-    resetField("password");
-    resetField("passwordConfirm");
-    toggleChangeHandler("password");
-  };
-
-  const changeNameHandler: SubmitHandler<UpdateInput> = async (data) => {
-    const inputValue = { name: data.name };
-    if (data.name !== currentName) {
-      if (!checkedDuplicate) {
-        setError("name", { message: "중복체크를 눌러주세요." });
-        return;
-      }
-
-      await changeMetaName(data.name);
-      patchUserMutation.mutate({ inputValue, userId });
-    } else {
-      toast("이전 닉네임과 동일합니다.", { theme: "failure", zIndex: 9999 });
-    }
-    toggleChangeHandler("name");
-  };
-
-  const deleteAuth = async () => {
-    const inputValue = { name: "탈퇴한 유저입니다.", avatar_url: "" };
-    if (await Confirm(<DeleteConfirmText />)) {
-      await deleteUser(userId);
-      patchUserMutation.mutate({ inputValue, userId });
-      await logout();
-      navigate("/");
-      toast("정상적으로 탈퇴되었습니다.", { theme: "plain", zIndex: 9999 });
-      if (prevProfileImageId !== "") {
-        void deleteImage(prevProfileImageId);
-      }
-    }
-  };
-
-  return (
-    <div className="relative mx-auto flex-column m-[60px] max-w-[1280px] w-[90%] sm:my-6">
-      <Title title="회원정보수정" isBorder={true} pathName="/mypage/update" />
-      <div className="flex w-full mt-10 sm:flex-col sm:gap-10 sm:contents-center">
-        <form onSubmit={handleSubmit(submitProfileImgHandler)} className="flex-column items-center w-[328px] gap-9">
-          <div className="relative w-[120px]">
-            <picture>
-              <source srcSet={currentProfileImg === "" ? defaultImgWebp : currentProfileImg} type="image/webp" />
-              <img
-                src={currentProfileImg === "" ? defaultImg : currentProfileImg}
-                alt="프로필이미지"
-                className="w-32 h-32 rounded-full"
-              />
-            </picture>
-            <div className="absolute bottom-0 flex items-center justify-center w-20 h-8 gap-2 -translate-x-1/2 bg-white border rounded-lg left-1/2 translate-y-1/4">
-              <label htmlFor="profileImgButton">
-                <img src={photoCamera} alt="이미지 등록" className="w-4 h-4 cursor-pointer" />
-              </label>
-              <input
-                id="profileImgButton"
-                type="file"
-                accept="image/png, image/jpeg, image/gif"
-                onChange={(event) => {
-                  imageSizeSaveHandler(event);
-                }}
-                className="hidden"
-              />
-              <div className="h-2 border bg-gray06" />
-              <img src={xmark} alt="닫기" onClick={resetImgFile} className="w-4 h-4 cursor-pointer" />
-            </div>
-          </div>
-          <p className="title-4">{`${currentName} 님`}</p>
-          {imageFile !== null && <button className="w-32 h-12 rounded-lg point-button body-3">이미지 변경하기</button>}
-        </form>
-        <div className="flex-column contents-center w-[624px] sm:w-full lg:w-full md:w-full">
-          <div className="w-full gap-6 flex-column sm:w-full">
-            <div className="gap-2 border-b pb-7 flex-column border-b-gray06 md:contents-center sm:contents-center sm:w-full">
-              <div className="flex gap-6">
-                <label htmlFor="nickname" className={LABEL_STYLE}>
-                  닉네임
-                </label>
-                <button
-                  id="nickname"
-                  type="button"
-                  onClick={() => {
-                    toggleChangeHandler("name");
-                  }}
-                  className="w-32 h-12 rounded-lg point-button body-3"
-                >
-                  변경
-                </button>
-              </div>
-
-              {isOpenToggle.name && (
-                <form onSubmit={handleSubmit(changeNameHandler)} className="flex-column">
-                  <div className="flex gap-6 w-[300px] mt-10 sm:flex-col sm:w-full">
-                    <input
-                      id={"name"}
-                      placeholder={"닉네임"}
-                      defaultValue={currentName}
-                      className="auth-input sm:w-full"
-                      {...register("name", {
-                        ...nameValid,
-                        onChange: () => {
-                          setCheckedDuplicate(false);
-                        },
-                      })}
-                    />
-                    <button
-                      type="button"
-                      onClick={duplicateCheck}
-                      className="w-32 h-12 rounded-lg gray-outline-button body-3 sm:w-full"
-                    >
-                      중복체크
-                    </button>
-                  </div>
-                  {checkedDuplicate ? (
-                    <p className={"h-10 w-full flex contents-center text-xs text-green-500 font-normal"}>
-                      사용 가능한 닉네임입니다.
-                    </p>
-                  ) : (
-                    <InvalidText className="justify-center" errorsMessage={errors.name?.message} size={40} />
-                  )}
-                  <div className="flex items-center justify-start gap-3 sm:justify-center">
-                    <button
-                      type="button"
-                      className="w-32 h-12 rounded-lg gray-outline-button body-3 sm:w-full"
-                      onClick={() => {
-                        toggleChangeHandler("name");
-                      }}
-                    >
-                      취소
-                    </button>
-                    <button className="w-32 h-12 rounded-lg point-button body-3 sm:w-full">수정</button>
-                  </div>
-                </form>
-              )}
-            </div>
-
-            <div className="gap-2 border-b pb-7 flex-column border-b-gray06 md:contents-center sm:contents-center sm:w-full">
-              <div className="flex gap-6 ">
-                <label htmlFor="password" className={LABEL_STYLE}>
-                  비밀번호
-                </label>
-                <button
-                  id="password"
-                  type="button"
-                  onClick={() => {
-                    toggleChangeHandler("password");
-                  }}
-                  className="w-32 h-12 rounded-lg point-button body-3"
-                >
-                  변경
-                </button>
-              </div>
-              {isOpenToggle.password && (
-                <form onSubmit={handleSubmit(changePasswordHandler)} className="sm:w-full">
-                  <div className="flex-column gap-2 w-[300px] sm:w-full mt-10">
-                    <div className="relative">
-                      <input
-                        placeholder="새 비밀번호"
-                        type={showPassword.password ? "text" : "password"}
-                        className="auth-input"
-                        {...register("password", {
-                          ...passwordValid(),
-                          validate: (_, formValue) =>
-                            formValue.password === formValue.passwordConfirm || "비밀번호가 일치하지 않습니다.",
-                        })}
-                      />
-                      <PasswordVisibleButton
-                        passwordType={"password"}
-                        isVisibleState={showPassword}
-                        setIsVisibleState={setShowPassword}
-                      />
-                    </div>
-                    <div className="relative">
-                      <input
-                        placeholder={"새 비밀번호 확인"}
-                        type={showPassword.passwordConfirm ?? false ? "text" : "password"}
-                        className="auth-input"
-                        {...register("passwordConfirm")}
-                      />
-                      <PasswordVisibleButton
-                        passwordType={"passwordConfirm"}
-                        isVisibleState={showPassword}
-                        setIsVisibleState={setShowPassword}
-                      />
-                    </div>
-                  </div>
-                  <InvalidText className={"justify-center"} errorsMessage={errors.password?.message} size={40} />
-                  <div className="flex items-center justify-start gap-3 sm:justify-center">
-                    <button
-                      type="button"
-                      className="w-32 h-12 rounded-lg gray-outline-button body-3 sm:w-full md:w-full"
-                      onClick={() => {
-                        toggleChangeHandler("password");
-                      }}
-                    >
-                      취소
-                    </button>
-                    <button className="w-32 h-12 rounded-lg point-button body-3 sm:w-full md:w-full">수정</button>
-                  </div>
-                </form>
-              )}
-            </div>
-          </div>
-          <div
-            className={`flex w-full py-6
-          lg:flex-column lg:gap-6 
-          md:w-full md:flex-col md:gap-6 
-          sm:w-full sm:flex-col sm:gap-6`}
-          >
-            <div className="flex items-center justify-center w-full gap-4 sm:w-full md:w-full">
-              <button
-                type="button"
-                className="flex w-[192px] h-12 bg-white rounded-lg contents-center gray-outline-button body-3 lg:w-60 sm:w-full md:w-full"
-                onClick={() => {
-                  navigate(-1);
-                }}
-              >
-                이전
-              </button>
-            </div>
-
-            <div
-              className={`absolute bottom-0 right-0 flex items-center gap-3 
-            lg:flex-col lg:static lg:contents-center
-            md:flex-col md:static md:contents-center
-            sm:flex-col sm:static sm:contents-center`}
-            >
-              <p className="body-3 text-gray02">더 이상 이용하지 않으시나요?</p>
-              <button
-                onClick={deleteAuth}
-                type="button"
-                className="w-[120px] h-12 border body-3 border-gray05 text-gray02 rounded-lg lg:w-60 md:w-full sm:w-full"
-              >
-                회원탈퇴
-              </button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default UpdateUser;
