+++ src/shared/Router.tsx
@@ -12,13 +12,13 @@ const Detail = lazy(async () => await import("../pages/Detail"));
 const Community = lazy(async () => await import("../pages/Community"));
 const Inquiry = lazy(async () => await import("../pages/Inquiry"));
 const Mypage = lazy(async () => await import("../pages/mypage/Mypage"));
-const UpdateUser = lazy(async () => await import("../pages/mypage/UpdateUser"));
+const UpdateUser = lazy(async () => await import("../pages/mypage/updateUser/UpdateUser"));
 const MyBookmark = lazy(async () => await import("../pages/mypage/MyBookmark"));
 const MyComment = lazy(async () => await import("../pages/mypage/MyComment"));
 const MyLike = lazy(async () => await import("../pages/mypage/MyLike"));
 const MyInquiry = lazy(async () => await import("../pages/mypage/MyInquiry"));
 const Login = lazy(async () => await import("../pages/Login"));
-const FindAuth = lazy(async () => await import("../pages/FindAuth"));
+const FindAuth = lazy(async () => await import("../pages/findAuth/FindAuth"));
 const UpdatePassword = lazy(async () => await import("../pages/UpdatePassword"));
 const AdminPage = lazy(async () => await import("../pages/adminPage/AdminPage"));
 const Signup = lazy(async () => await import("../pages/Signup"));
